// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/daily-micronutrient-summary-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DailyMicronutrientHttpResponse } from './daily-micronutrient-http-response';
import { HealthAreaScoresHttpResponse } from './health-area-scores-http-response';

export interface DailyMicronutrientSummaryHttpResponse {
    date?: string;
    healthScores?: HealthAreaScoresHttpResponse | null;
    linkedProductCount?: number;
    nutrients?: Array<DailyMicronutrientHttpResponse>;
    totalProductCount?: number;
}