// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/health-area-scores-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { HealthAreaScoreHttpResponse } from './health-area-score-http-response';

export interface HealthAreaScoresHttpResponse {
    antioxidant?: HealthAreaScoreHttpResponse;
    bone?: HealthAreaScoreHttpResponse;
    energy?: HealthAreaScoreHttpResponse;
    heart?: HealthAreaScoreHttpResponse;
    immune?: HealthAreaScoreHttpResponse;
}