// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/usda-food-detail-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { HealthAreaScoresHttpResponse } from './health-area-scores-http-response';
import { MicronutrientHttpResponse } from './micronutrient-http-response';
import { UsdaFoodPortionHttpResponse } from './usda-food-portion-http-response';

export interface UsdaFoodDetailHttpResponse {
    description?: string;
    fdcId?: number;
    foodCategory?: string | null;
    healthScores?: HealthAreaScoresHttpResponse | null;
    nutrients?: Array<MicronutrientHttpResponse>;
    portions?: Array<UsdaFoodPortionHttpResponse>;
}